Move context related stuff to its own package.
