+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/OneTimeLink.kt
@@ -1,5 +1,7 @@
 package org.vaccineimpact.api.app
 
+import org.vaccineimpact.api.app.context.ActionContext
+import org.vaccineimpact.api.app.context.OneTimeLinkActionContext
 import org.vaccineimpact.api.serialization.Deserializer
 import org.vaccineimpact.api.models.helpers.OneTimeAction
 import org.vaccineimpact.api.app.controllers.MontaguControllers
