+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/app_start/Controller.kt
@@ -1,5 +1,5 @@
 package org.vaccineimpact.api.app.app_start
 
-import org.vaccineimpact.api.app.ActionContext
+import org.vaccineimpact.api.app.context.ActionContext
 
 abstract class Controller(val context: ActionContext)
