+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/context/ActionContext.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api.app
+package org.vaccineimpact.api.app.context
 
 import org.pac4j.core.profile.CommonProfile
 import org.vaccineimpact.api.models.permissions.PermissionSet
