+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/filters/ScenarioFilterParameters.kt
@@ -1,6 +1,6 @@
 package org.vaccineimpact.api.app.filters
 
-import org.vaccineimpact.api.app.ActionContext
+import org.vaccineimpact.api.app.context.ActionContext
 
 data class ScenarioFilterParameters(val scenarioId: String? = null,
                                     val disease: String? = null)
