+++ src/app/src/test/kotlin/org/vaccineimpact/api/tests/controllers/AuthenticationControllerTests.kt
@@ -3,7 +3,7 @@ package org.vaccineimpact.api.tests.controllers
 import com.nhaarman.mockito_kotlin.*
 import org.junit.Test
 import org.pac4j.core.profile.CommonProfile
-import org.vaccineimpact.api.app.ActionContext
+import org.vaccineimpact.api.app.context.ActionContext
 import org.vaccineimpact.api.app.FormHelpers
 import org.vaccineimpact.api.app.HTMLForm
 import org.vaccineimpact.api.app.controllers.AuthenticationController
