+++ src/app/src/test/kotlin/org/vaccineimpact/api/tests/controllers/EndpointTests.kt
@@ -2,7 +2,7 @@ package org.vaccineimpact.api.tests.controllers
 
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
-import org.vaccineimpact.api.app.ActionContext
+import org.vaccineimpact.api.app.context.ActionContext
 import org.vaccineimpact.api.app.controllers.endpoints.basicEndpoint
 import org.vaccineimpact.api.test_helpers.MontaguTests
