+++ src/app/src/test/kotlin/org/vaccineimpact/api/tests/controllers/userController/GetUserTests.kt
@@ -4,7 +4,7 @@ import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions
 import org.junit.Test
-import org.vaccineimpact.api.app.ActionContext
+import org.vaccineimpact.api.app.context.ActionContext
 import org.vaccineimpact.api.app.controllers.UserController
 import org.vaccineimpact.api.app.repositories.UserRepository
 import org.vaccineimpact.api.models.User
